pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wildlife-tools"
version = "0.0.2"
authors = [
  { name="Vojtěch Čermák", email="cermak.vojtech@seznam.cz" },
]
maintainers = [
  { name="Vojtěch Čermák", email="cermak.vojtech@seznam.cz" },
]
description = "Tools for using wildlife re-identification datasets."
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
keywords = [
    "wildlife",
    "re-identification",
    "datasets",
    "machine learning"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies=[
          'torch>=2.0.1',
          'timm>=0.9.2',
          'numpy>=1.19.4',
          'pandas>=1.1.4',
          'tqdm>=4.62.3',
          'opencv-python>=4.5.5.62',
          'Pillow>=7.1.2',
          'scikit-learn>=1.0.1',
          'faiss-gpu>=1.7.2',
]

[project.urls]
"Homepage" = "https://github.com/WildlifeDatasets/wildlife-tools"
"Documentation" = "https://wildlifedatasets.github.io/wildlife-tools/"
"Bug Tracker" = "https://github.com/WildlifeDatasets/wildlife-tools/issues"


[tool.setuptools]
packages = [
  "wildlife_tools",
  "wildlife_tools.data",
  "wildlife_tools.evaluation",
  "wildlife_tools.features",
  "wildlife_tools.similarity",
  "wildlife_tools.train",
]

